Much improved code for extracting the coordinates from SVG path data.
It's also based on LaTeX documentclass article again, not relying on cropped PDFs anymore, but rather calculating the translation of the node and offsetting it.

Inkscape changes that path data once you move a node around. It then contains coordinates that are relative instead of absolute.
There are also some path commands that only have one coordinate part (horizontal and vertical movement/lines). These are now considered as well. This made the placement of new and edited nodes more exact.

--HG--
branch : tu_dresden