Bug with moved nodes fixed.
Inkscape converts paths to relative format when a node is moved around. At the same time it makes use of an optimization that skips command letters in the path, when the same command type happens multiple times in a row.

Example:  v 2 v 1 v 3  would draw three vertical lines in a row and can be simplified to v 2 1 3. This previously upset the algorithm I used to calculate the size of the node.
(See: http://www.w3.org/TR/SVG/paths.html#PathData)

Updated installer.