Modified: Documentation structure.
The base directory contains a file README.md the syntax of which is compatible
to the bitbucket webinterface. The directory docs contains a file README.md
which is adapted to the pandoc document generator in order to typeset a pretty
pdf and html documentation (Call build_doc.bat to build the pdf and html
documentation using pandoc).

ToDo: Generate either the pandoc .md or the bitbucket .md file from the other
one in order to avoid text doubling...