Fixed: Bad mistake in installer.nsi preventing full user installation
Registry-Entries for PyGTK are written into HKCU instead of HKLM now. If they
are requested to be written into HKLM and RequestExecutionLevel=user the
write process silently (!) fails. Now we can be sure that a) the textext
files go into the correct AppData directory (i.e., the user's one) and b) PyGTK
still can find the required registry-strings.