Added: Installer scripts for PyGTK for Inkscape
In Version 0.6 of TexText the required PyGTK-package will be provided by an
own installer. The reason for this is the fact that the PyGTK all in one
package which can be downloaded from the PyGTK site is not able to install
its contents into a 64-bit installation of Inkscape (independently of the
fact that registry keys for PythonCore are properly set).