Added detection of wrong ghostscript version.
If ghostscript > 9.20 is installed on the system pstoedit fails due to the
deprecated DELAYBIND option passed while invoking the gs command (see
issue #48 and issue #50).

An indiction for this error is that a svg file of size 0 bytes is created. This
is checked and a corresponding message is sent to the log.

Note that no exception is thrown from subprocess.communicate if this error
occurs. Hence, we have to check it that way.

Note also, that some Linux distros fixed gs and/ or pstoedit so a direct check
of the gs version is no solution.

--HG--
branch : hotfix/detect_wrong_gs_version