Fixed: Typesetter import error when LaTeX compilation failed
Under Windows the extension subfolder is not part of the path anymore in
Inkscape 0.92.2. Hence, the import of the Typesetter module fails in
LatexConverterBase.parse_pdf_log(...) where we are in the temp directory. As
a workaround the import is done at the top of the file where we can be sure to
be in the extension subfolder. There are no more imports which are affected by
this problem. I prefer this as a workaround instead of messing around with
the system path.

--HG--
branch : releases/0.7