Store TexText version and used converters in node
For future improvements it is helpful to have information about the
TexText version as well as about the TeX converter (LaTeX, XeTeX,
LuaTex) and the pdf converter (pstoedit, pdf2svg) used for creating the
node.

Since this feature is required for the pdf2svg support as well as for
the font scale hotfix it is introduced via this separate commit.