Fixed vertical flipping of nodes if pdf converter is changed
When a node which has been previously created using pstoedit is
converted with pdf2svg (and vice versa) the output is vertically flipped.
This fix ensures that this does not happen anymore. To ensure this
happens in a clean way the whole alignment procedure has been moved from
TexText.do_convert into to SvgElement which seems to be quite natural.