After conversion strokes with "none" color are set tpo zero width
If this is not done problems occur when an SVG node is colorized after
creation via Inkscape. Reason is the following:

pdf2svg produces a mixture of paths which have a
fill color but no stroke color (and no stroke-width defined) and of paths which have a stroke color
but no fill color. To colorize nodes of such mixtures in Inkscape one
needs to assign the fill color and the stroke color simultaneously.

Since filled paths with stroke color "none" and no stroke-width defined
have a stroke width of 1 by default, the filled paths look very ugly
after a stroke color has been assigned because a colorized stroke is visible around them.

By setting the stroke width to 0 for paths which are filled by pdf2svg
with stroke color none, this can be avoided and one has the same
behavior as when editing pstoedit nodes.