Removed stand alone pdf user guide/readme
The stand alone readme shipped with TexText basically contained the same
stuff as the Readme.md of the repository. Especially, it does not make
sense to repeat all the installation stuff when the extension has
already been installed. The wiki presents the required information now.
The installers have been adapted.

Resolves #9