Doc: Updated for installation of RC1
Doc ensures that the proper Inkscape RC1 package for Linux is downloaded (they
are currently not available on the regular inkscape web page)

Later this will be modifed such that the most recent version can be
downloaded as for Windows on the Inkscape web page. Requires some
additional work due to screeshots that have to be taken.