Add "stroke-width=0pt" to SVG elements with "stroke=none"
For each element which has the style attribute "stroke"
set to "none" a style attribute "stroke-width" with
value "0" is added. This ensures that when colorizing
the node later in inkscape by setting the node and
stroke colors letters do not become bold (letters have
"stroke" set to "none" but e.g. horizontal lines in
fraction bars and square roots are only affected by
stroke colors so for full colorization of a node you
need to set the fill as well as the stroke color!).