Respect Inkscape's 1.2 interpretation of user units as px
Since Inkscape 1.2 user units are always interpreted as px, not in the document units as before. Hence, behaviour of the method unittouu has changed as it always returns the conversion to px. Especially, the center attribute of svg.namedview is now returned in px and not in the document units. This is addressed by this fix for both, Inkscape 1.0/1.1 and Inkscape 1.2.

Resolves #338