Show errmsg if TexText fails with unhandled exception
If TexText is not able to handle errors
by itself (i.e. is not able to show an
error message) the error is piped to
stderr. Up to Inkscape 1.3 these errors
were not processed correctly by
Inkscape. Since version 1.4. it is
possible to show such errors by
passing the show-stderr option in
the inx file.

This also means that minimum Inkscape
requirement is 1.4 now since we make
use of the show-stderr option in the
inx file.